% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admm_PLN.R
\name{admmPLN}
\alias{admmPLN}
\title{admmPLN}
\usage{
admmPLN(Xdata, mu.ini = NULL, A.ini = NULL, Z.ini = NULL, U.ini = NULL,
  lambda, rho, scale = NULL, max.iter = 300, max.iter.estA = 3,
  eps.abs = 1e-06, eps.rel = 0.001, use.diag = TRUE,
  penalize.diag = FALSE, approx.extended = FALSE, weightBydegree = TRUE,
  verbose = FALSE)
}
\arguments{
\item{Xdata}{a nxp counts data matrix}

\item{mu.ini}{a initial value of means for p genes}

\item{A.ini}{a initial value of A matrix, pxp}

\item{Z.ini}{a initial value of Z matrix, pxp}

\item{U.ini}{a initial value of U matrix, pxp}

\item{lambda}{a tuning parameter}

\item{rho}{a Lagrangian parameter}

\item{scale}{a nxp matrix of scale K}

\item{max.iter}{the maximum number of iterations, the default is 300.}

\item{max.iter.estA}{the maximum number of iterations for estA()}

\item{eps.abs}{a threshold for ADMM convergence}

\item{eps.rel}{a threshold for ADMM convergence}

\item{use.diag}{whether we use diagonal terms only to facilitate computation, default is TRUE.}

\item{penalize.diag}{whether we penalize diagonal terms of A or not, default is FALSE.}

\item{approx.extended}{whether we extend Laplace's approximation or not, default is FALSE.}

\item{weightBydegree}{whether we use weights depending on degrees, default is TRUE.}
}
\value{
a list of estimated A, Z, U and mu
}
\description{
This function implements ADMM to estimate A based on Poisson Log-normal model.
}
\examples{
res <- admmPLN(simX,lambda=0.2,rho=0.5)
}
