% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_U.R
\name{estU}
\alias{estU}
\title{estU}
\usage{
estU(A, Z, U)
}
\arguments{
\item{A}{a pxp matrix}

\item{Z}{a pxp matrix}

\item{U}{a pxp matrix}
}
\value{
updated U matrix
}
\description{
This function updates U for given A and Z
}
