% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_Z.R
\name{estZ}
\alias{estZ}
\title{estZ}
\usage{
estZ(Z, A, U, lambda, rho, weightBydegree = TRUE)
}
\arguments{
\item{Z}{a pxp matrix}

\item{A}{a pxp matrix}

\item{U}{a pxp matrix}

\item{lambda}{a tuning parameter}

\item{rho}{a Lagrangian parameter}

\item{weightBydegree}{whether we use weights depending on degrees or not, default is TRUE}
}
\value{
updated Z matrix
}
\description{
This function updates Z given A, U
}
