% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_eta.R
\name{etaUpdate}
\alias{etaUpdate}
\title{etaUpdate}
\usage{
etaUpdate(mu, A, diag.A, Xn, eta.old = log(Xn + 1), scale.n = NULL,
  eps = 1e-06, max.iter = 100, use.diag = TRUE)
}
\arguments{
\item{mu}{a mean of p genes, px1 vector}

\item{A}{an inverse of covariance matrix, pxp matrix}

\item{Xn}{counts for a sample, px1}

\item{eta.old}{initial estimate of eta}

\item{scale.n}{a K vector for a sample, px1 vector}

\item{eps}{a criterion for convergence where d(eta.old, eta.new) < eps}

\item{max.iter}{the maximum number of iterations}
}
\value{
a list of eta.star and sigma.star
}
\description{
This function computes eta.star for given mu and A to maximize the likelihood
}
