% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted_edge.R
\name{fittedEdge}
\alias{fittedEdge}
\title{fittedEdge}
\usage{
fittedEdge(A, eps)
}
\arguments{
\item{A}{a concentration matrix, pxp}

\item{eps}{a threshold for partial correlation}
}
\value{
a pxp indication matrix
}
\description{
This function computes a indication matrix for edges,
i.e. (i,j) = 1 if the partial correlation between gene i and j is larger than eps
           = 0 otherwise
}
