% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mme_PLN.R
\name{mmePLN}
\alias{mmePLN}
\title{mmePLN - Moment method estimate}
\usage{
mmePLN(Xdata, scale = NULL)
}
\arguments{
\item{Xdata}{a matrix of gene counts nxp where n is the number of samples and p is the number of genes.}

\item{scale}{K, default is NULL}
}
\value{
MME of mu and sigma
}
\description{
This function computes MME of mu and sigma based on poisson log-normal distribution
}
