\name{obs_mat}
\alias{obs_mat}
\docType{data}
\title{
A simulation data for gene regulatory network inference.
}
\description{
A simulation single-cell RNA sequence data with 100 cells and 50 genes generated by poisson log-normal model.
}
\usage{data("realdatamatrix")}
\format{
  A data frame with 100 rows and 50 columns.
}

\examples{
data("obs_mat")
}
\keyword{datasets}
