\name{realdatamatrix}
\alias{realdatamatrix}
\docType{data}
\title{
A real data for gene regulatory network inference.
}
\description{
A single-cell RNA sequence data about Interferon β-Stimulated PBMCs in CD14 monocytes celltype with 2147 cells and 200 genes.
}
\usage{data("realdatamatrix")}
\format{
  A data frame with 2147 rows and 200 columns.
}

\examples{
data("realdatamatrix")
}
\keyword{datasets}
