% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_soft.R
\name{thresholdS}
\alias{thresholdS}
\title{thresholdS}
\usage{
thresholdS(A, B)
}
\arguments{
\item{A}{a pxp matrix}

\item{B}{a pxp matrix}
}
\value{
a pxp matrix
}
\description{
This function computes soft thresholding
ST(Aij, Bij) = Aij - Bij if Aij>0 & Aij-Bij>0,
               Aij + Bij if Aij<0 & Aij+Bij<0,
               0 otherwise
}
