% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_A.R
\name{updateA}
\alias{updateA}
\title{updateA, a sub-function of estA()}
\usage{
updateA(A, gr, alpha)
}
\arguments{
\item{A}{a pxp matrix}

\item{gr}{the gradient}

\item{alpha}{a step size}
}
\value{
a list of A and alpha
}
\description{
This function updates A as a sub-step of function, estA() and
checks whether A is positive definite
}
